[iOS] [12369] working on Notifications
