+++ OpenStack Summit/OpenStack Summit/PushNotificationManager.swift
@@ -121,7 +121,7 @@ public final class PushNotificationManager: NSObject, NSFetchedResultsController
             if teamMessage.from.identifier != store.authenticatedMember?.identifier {
                 
                 let userNotification = UILocalNotification()
-                userNotification.userInfo = [UserNotificationUserInfo.topic.rawValue: PushNotificationTopic.team(teamMessage.team.identifier).rawValue]
+                userNotification.userInfo = [UserNotificationUserInfo.topic.rawValue: Notification.Topic.team(teamMessage.team.identifier).rawValue]
                 userNotification.alertTitle = "\(teamMessageNotification.from.firstName) \(teamMessageNotification.from.lastName)"
                 userNotification.alertBody = teamMessageNotification.body
                 userNotification.fireDate = NSDate()
@@ -164,7 +164,7 @@ public final class PushNotificationManager: NSObject, NSFetchedResultsController
             case .reply:
                 
                 guard let topicString = notification.userInfo?[UserNotificationUserInfo.topic.rawValue] as? String,
-                    let topic = PushNotificationTopic(rawValue: topicString),
+                    let topic = Notification.Topic(rawValue: topicString),
                     case let .team(team) = topic
                     else { completion(); return }
                 
@@ -214,7 +214,7 @@ public final class PushNotificationManager: NSObject, NSFetchedResultsController
     }
     
     @inline(__always)
-    private func subscribe(to topic: PushNotificationTopic) {
+    private func subscribe(to topic: Notification.Topic) {
         
         FIRMessaging.messaging().subscribeToTopic(topic.rawValue)
         
@@ -222,7 +222,7 @@ public final class PushNotificationManager: NSObject, NSFetchedResultsController
     }
     
     @inline(__always)
-    private func unsubscribe(from topic: PushNotificationTopic) {
+    private func unsubscribe(from topic: Notification.Topic) {
         
         FIRMessaging.messaging().unsubscribeFromTopic(topic.rawValue)
         
@@ -308,7 +308,7 @@ public final class PushNotificationManager: NSObject, NSFetchedResultsController
         
         let identifier = (anObject as! Entity).identifier
         
-        let topic: PushNotificationTopic
+        let topic: Notification.Topic
         
         if controller === teamsFetchedResultsController {
             
@@ -389,146 +389,6 @@ public protocol PushNotification {
     init?(pushNotification: [String: String])
 }
 
-public enum PushNotificationTopic {
-    
-    case summit(Identifier)
-    case team(Identifier)
-    case member(Identifier)
-    case event(Identifier)
-    case attendees
-    case speakers
-    case everyone
-    
-    public var identifier: Identifier? {
-        
-        switch self {
-        case let .summit(identifier): return identifier
-        case let .team(identifier): return identifier
-        case let .member(identifier): return identifier
-        case let .event(identifier): return identifier
-        default: return nil
-        }
-    }
-    
-    private enum Prefix: String {
-        
-        case summit, member, event, team, attendees, speakers, everyone
-        
-        init(_ topic: PushNotificationTopic) {
-            
-            switch topic {
-                
-            case .summit:       self = .summit
-            case .team:         self = .team
-            case .member:       self = .member
-            case .event:        self = .event
-            case .attendees:    self = .attendees
-            case .speakers:     self = .speakers
-            case .everyone:     self = .everyone
-            }
-        }
-    }
-    
-    public init?(rawValue: String) {
-        
-        struct Cache {
-            static var prefixRegularExpressions = [Prefix: NSRegularExpression]()
-        }
-        
-        func parseIdentifier(prefix: Prefix) -> Identifier? {
-            
-            let prefixString = "/topics/" + prefix.rawValue + "_"
-            
-            guard rawValue.containsString(prefixString) else { return nil }
-            
-            // get regex
-            
-            let regularExpression: NSRegularExpression
-            
-            if let cached = Cache.prefixRegularExpressions[prefix] {
-                
-                regularExpression = cached
-                
-            } else {
-                
-                regularExpression = try! NSRegularExpression(pattern: prefixString + "(\\d+)", options: [])
-                
-                Cache.prefixRegularExpressions[prefix] = regularExpression
-            }
-            
-            // run regex
-            
-            guard let match = regularExpression.firstMatchInString(rawValue, options: [], range: NSMakeRange(0, (rawValue as NSString).length))
-                where match.numberOfRanges == 2
-                else { return nil }
-            
-            let matchString = (rawValue as NSString).substringWithRange(match.range)
-            
-            guard matchString == rawValue
-                else { return nil }
-            
-            let captureGroup = (rawValue as NSString).substringWithRange(match.rangeAtIndex(1))
-            
-            guard let identifier = Int(captureGroup)
-                else { return nil }
-            
-            return identifier
-        }
-        
-        func parseCollection() -> PushNotificationTopic? {
-            
-            guard let prefix = Prefix(rawValue: rawValue)
-                else { return nil }
-            
-            switch prefix {
-            case .attendees: return .attendees
-            case .everyone: return .everyone
-            case .speakers: return .speakers
-            default: return nil
-            }
-        }
-        
-        // main `if-else` parsing statement
-        
-        if let collectionTopic = parseCollection() {
-            
-            self = collectionTopic
-            
-        } else if let identifier = parseIdentifier(.summit) {
-            
-            self = .summit(identifier)
-            
-        } else if let identifier = parseIdentifier(.member) {
-            
-            self = .member(identifier)
-            
-        } else if let identifier = parseIdentifier(.event) {
-            
-            self = .event(identifier)
-            
-        } else if let identifier = parseIdentifier(.team) {
-            
-            self = .team(identifier)
-            
-        } else {
-            
-            return nil
-        }
-    }
-    
-    public var rawValue: String {
-        
-        var stringValue = "/topics/" + Prefix(self).rawValue
-        
-        if let identifier = self.identifier {
-            
-            stringValue += "_\(identifier)"
-        }
-        
-        return stringValue
-    }
-}
-
 public struct TeamMessageNotification: PushNotification {
     
     private enum Key: String {
@@ -551,7 +411,7 @@ public struct TeamMessageNotification: PushNotification {
     public init?(pushNotification: [String: String]) {
         
         guard let topicString = pushNotification[Key.from.rawValue],
-            let topic = PushNotificationTopic(rawValue: topicString),
+            let topic = Notification.Topic(rawValue: topicString),
             let typeString = pushNotification[Key.type.rawValue],
             let type = PushNotificationType(rawValue: typeString),
             case let .team(team) = topic,
@@ -590,17 +450,6 @@ public extension TeamMessage {
 
 public struct GeneralNotification: PushNotification {
     
-    public enum Channel: String {
-        
-        case everyone = "EVERYONE"
-        case speakers = "SPEAKERS"
-        case attendess = "ATTENDEES"
-        case members = "MEMBERS"
-        case summit = "SUMMIT"
-        case event = "EVENT"
-        case group = "GROUP"
-    }
-    
     private enum Key: String {
         
         case from, id, type, body, summit_id, channel, created_at, event_id, title
@@ -614,18 +463,18 @@ public struct GeneralNotification: PushNotification {
     
     public let created: Date
     
-    public let from: PushNotificationTopic
+    public let from: Notification.Topic
     
     public let summit: Identifier
     
-    public let channel: Channel
+    public let channel: Notification.Channel
     
     public let event: (identifier: Identifier, title: String)?
     
     public init?(pushNotification: [String: String]) {
         
         guard let topicString = pushNotification[Key.from.rawValue],
-            let topic = PushNotificationTopic(rawValue: topicString),
+            let topic = Notification.Topic(rawValue: topicString),
             let typeString = pushNotification[Key.type.rawValue],
             let type = PushNotificationType(rawValue: typeString),
             let identifierString = pushNotification[Key.id.rawValue],
@@ -636,7 +485,7 @@ public struct GeneralNotification: PushNotification {
             let summitIDString = pushNotification[Key.summit_id.rawValue],
             let summitID = Int(summitIDString),
             let channelString = pushNotification[Key.channel.rawValue],
-            let channel = Channel(rawValue: channelString)
+            let channel = Notification.Channel(rawValue: channelString)
             where type == self.dynamicType.type
             else { return nil }
